Fix comment tutorial not showing up on mobile
