+++ Apps/WebClient/src/ClientApp/src/components/timeline/entryCard/CommentSectionComponent.vue
@@ -22,15 +22,23 @@ import { ILogger } from "@/services/interfaces";
     },
 })
 export default class CommentSectionComponent extends Vue {
-    @Prop() parentEntry!: TimelineEntry;
-    @Prop({ default: false }) isMobileDetails!: boolean;
+    @Prop()
+    parentEntry!: TimelineEntry;
 
-    @Getter("user", { namespace: "user" }) user!: User;
+    @Prop({ default: false })
+    isMobileDetails!: boolean;
+
+    @Prop({ default: false })
+    visible!: boolean;
+
     @Action("updateComment", { namespace: "comment" })
     updateComment!: (params: {
         hdid: string;
         comment: UserComment;
     }) => Promise<UserComment>;
+
+    @Getter("user", { namespace: "user" })
+    user!: User;
 
     private logger!: ILogger;
     private showComments = false;
@@ -169,6 +177,8 @@ export default class CommentSectionComponent extends Vue {
                     'fixed-bottom p-3 comment-input': isMobileDetails,
                 }"
                 :comment="newComment"
+                :is-mobile-details="isMobileDetails"
+                :visible="visible"
                 @on-comment-added="onAdd"
             ></AddComment>
         </b-col>
